use papaparse for parsing json to csv
Canvas seems unhappy with putting  ISO dates in quotes,
so rather than dumb json to csv parsing where every
value is put in quotes and separated by a comma, we're
changing to a package for parsing. papaparse only
includes quotes if needed.

This change also caused the csv output to change, so
a number of tests needed to be updated. Where possible,
we've changed the tests to check for a match against
inline snapshots.